agents/tool/requirements.txt
# Tool agent dependencies
# Shared dependencies are in agents/shared/requirements.txt
strands-agents[a2a]>=1.20.0
requests>=2.32.5

